shuffle_soc_factory/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
urllib3>=2.0.0
